// collabarative-design-tool-main/src/utils.ts
import {
  LayerType,
  Side,
  type Camera,
  type Color,
  type PathLayer,
  type Point,
  type XYWH,
} from "./types";

export function rgbToHex(color: Color) {
  const { r, g, b } = color;
  const toHex = (n: number) => n.toString(16).padStart(2, "0");
  return `#${toHex(r)}${toHex(g)}${toHex(b)}`;
}

export function penPointToPathLayer(
  points: number[][],
  color: Color,
): PathLayer {
  let left = Number.POSITIVE_INFINITY;
  let right = Number.NEGATIVE_INFINITY;
  let top = Number.POSITIVE_INFINITY;
  let bottom = Number.NEGATIVE_INFINITY;

  for (const point of points) {
    const [x, y] = point;
    if (x === undefined || y === undefined) continue;
    if (left > x) {
      left = x;
    }
    if (right < x) {
      right = x;
    }
    if (top > y) {
      top = y;
    }
    if (bottom < y) {
      bottom = y;
    }
  }

  return {
    type: LayerType.Path,
    x: left,
    y: top,
    width: right - left,
    height: bottom - top,
    fill: color,
    stroke: color,
    opacity: 100,
    points: points
      .filter(
        (point): point is [number, number, number] =>
          point[0] !== undefined &&
          point[1] !== undefined &&
          point[2] !== undefined,
      )
      .map(([x, y, pressure]) => [x - left, y - top, pressure]),
  };
}

export function getSvgPathFromStroke(stroke: number[][]) {
  if (!stroke.length) return "";

  const d = stroke.reduce(
    (acc, [x0, y0], i, arr) => {
      const nextPoint = arr[(i + 1) % arr.length];

      if (!nextPoint) return acc;

      const [x1, y1] = nextPoint;
      acc.push(x0!, y0!, (x0! + x1!) / 2, (y0! + y1!) / 2);
      return acc;
    },
    ["M", ...(stroke[0] ?? []), "Q"],
  );

  d.push("Z");
  return d.join(" ");
}

// export const pointerEventToCanvasPoint = (
//   e: React.PointerEvent,
//   camera: Camera,
// ): Point => {
//   return {
//     x: Math.round(e.clientX) - camera.x,
//     y: Math.round(e.clientY) - camera.y,
//   };
// };

export function pointerEventToCanvasPoint(e: React.PointerEvent, camera: Camera) {
  const svg = (e.currentTarget as SVGSVGElement);
  const rect = svg.getBoundingClientRect();

  const x = (e.clientX - rect.left - camera.x) / camera.zoom;
  const y = (e.clientY - rect.top - camera.y) / camera.zoom;

  return { x, y };
}


export function resizeBounds(bounds: XYWH, corner: Side, point: Point): XYWH {
  const result = {
    x: bounds.x,
    y: bounds.y,
    width: bounds.width,
    height: bounds.height,
  };

  if (corner === Side.Left || (corner & Side.Left) !== 0) {
    result.x = Math.min(point.x, bounds.x + bounds.width);
    result.width = Math.abs(bounds.x + bounds.width - point.x);
  }
  if (corner === Side.Right || (corner & Side.Right) !== 0) {
    result.x = Math.min(point.x, bounds.x);
    result.width = Math.abs(point.x - bounds.x);
  }
  if (corner === Side.Top || (corner & Side.Top) !== 0) {
    result.y = Math.min(point.y, bounds.y + bounds.height);
    result.height = Math.abs(bounds.y + bounds.height - point.y);
  }
  if (corner === Side.Bottom || (corner & Side.Bottom) !== 0) {
    result.y = Math.min(point.y, bounds.y);
    result.height = Math.abs(point.y - bounds.y);
  }

  return result;
}


